dsv4: audit follow-ups — int64 grid index in hc kernels, reuse set_input staging buffers
Full-branch code audit found no correctness bugs (the prior 12-fix audit
caught the real ones); these are the two perf/hygiene items it surfaced:

- kernel_dsv4_hc_expand / hc_weighted_sum: compute the flat thread index
  as int64 instead of int32. Not reachable at 16K context (max n_elem
  ~2.7e8, 8x under 2^31) but removes the overflow footgun for large ubatches.
- dsv4_graph_inputs::set_input: stage ivec/mask data in reused member
  buffers (ivec_buf/mask_buf) instead of a fresh std::vector per ivec and
  per mask every decode token. The input object is shared across decode
  tokens of a reused graph, so after the first token there is no per-token
  heap alloc on the hot path (was hundreds of malloc/free per token across
  compress layers). Greedy output identical, reused 47/47.

+++ ggml/src/ggml-cuda/dsv4.cu
@@ -255,8 +255,8 @@ static __global__ void kernel_dsv4_hc_expand(
         const char * comb,
         char * dst) {
     const int64_t n_elem = args.n_embd * args.n_hc * args.n_tokens;
-    const int gid = blockIdx.x * blockDim.x + threadIdx.x;
-    if ((int64_t) gid >= n_elem) {
+    const int64_t gid = (int64_t) blockIdx.x * blockDim.x + threadIdx.x;
+    if (gid >= n_elem) {
         return;
     }
 
@@ -427,8 +427,8 @@ static __global__ void kernel_dsv4_hc_weighted_sum(
         const char * weights,
         char * dst) {
     const int64_t n_elem = args.n_embd * args.n_tokens;
-    const int gid = blockIdx.x * blockDim.x + threadIdx.x;
-    if ((int64_t) gid >= n_elem) {
+    const int64_t gid = (int64_t) blockIdx.x * blockDim.x + threadIdx.x;
+    if (gid >= n_elem) {
         return;
     }
